requirements.txt
discord.py>=2.3.2
python-dotenv==1.0.0
pymongo>=4.5.0
dnspython>=2.3.0
httpx>=0.24.1
Pillow>=10.0.0
numpy>=1.24.0
emoji>=2.2.0
python-dateutil>=2.8.2
pytz>=2023.3
